ANE-accelerated VAE decoder for Sana image generation (3x speedup)
Adds Apple Neural Engine acceleration for the Sana VAE decoder,
reducing 512x512 image generation from ~16s to ~5s (4 diffusion steps).

VAE decode specifically goes from ~10s on CPU to <0.3s on ANE.

CoreML conversion (python/convert_sana_coreml.py):
- Patches movedim, repeat_interleave, RMSNorm, and attention
  reshapes for coremltools trace compatibility
- Traces in float32 for fast CPU tracing, CoreML handles
  device-specific compilation on first run
- Uses compute_precision=FLOAT32 to avoid FP16 overflow in
  deep RMSNorm chains

C++ runtime integration:
- SanaModel loads vae_decoder.mlpackage via NPUEncoder at init
- decode_latents() uses ANE when available, with shape validation
  to fall back to CPU for unsupported resolutions
- npu_ane.mm handles float32 MLMultiArray output conversion

Kernel optimizations:
- Linear attention uses cblas_sgemm (Accelerate) with parallel
  head processing instead of scalar triple-nested loops
- Depthwise conv2d NEON fast path for stride=1 dilation=1

Also: load HF models in float16 during weight conversion to
reduce memory, default run_txt2img.py to 512x512

+++ python/convert_sana.py
@@ -63,7 +63,7 @@ def load_diffusers_component(model_cls, model_id, subfolder, variant: Optional[s
     attempts = [a for a in attempts if not (a in seen or seen.add(a))]
     last_error = None
     for candidate in attempts:
-        kwargs = dict(subfolder=subfolder, torch_dtype=torch.float32)
+        kwargs = dict(subfolder=subfolder, torch_dtype=torch.float16)
         if candidate is not None:
             kwargs["variant"] = candidate
         try:
@@ -95,7 +95,7 @@ def convert_text_encoder(output_dir, precision="INT8", model_id="google/gemma-2-
     from transformers import AutoModelForCausalLM, AutoTokenizer
     hf_token = os.environ.get("HF_TOKEN") or os.environ.get("HUGGINGFACE_HUB_TOKEN")
 
-    model = AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=torch.float32, token=hf_token)
+    model = AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=torch.float16, token=hf_token)
     
     te_dir = Path(output_dir) / "text_encoder"
     te_dir.mkdir(parents=True, exist_ok=True)
